Fix: Change Pv1Voltage/Pv2Voltage to double type in DeviceRealTimeData
- Fix build error: Cannot implicitly convert type double to int
- Update Pv1Voltage from int to double
- Update Pv2Voltage from int? to double?

+++ LumenTreeInfo.API/Models/DeviceRealTimeData.cs
@@ -29,8 +29,8 @@ public class DeviceRealTimeData(string deviceId)
     // PV (Solar Panel) data
     public int Pv1Power { get; set; }
     public int? Pv2Power { get; set; }
-    public int Pv1Voltage { get; set; }
-    public int? Pv2Voltage { get; set; }
+    public double Pv1Voltage { get; set; }
+    public double? Pv2Voltage { get; set; }
     public int PvTotalPower { get; set; }
     
     // Battery cell voltages
